feat: Merge fix--refacotr-getcontent-to-take-optional-param-dispatch
